adjust external API
